#[codec(skip)] will do, and combine map_key stuff into storage_address since it's all specific to that
